+++ crates/nostr/CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Added
 
 - Add nip47 holdinvoice methods and notification (https://github.com/rust-nostr/nostr/pull/1019)
+- Add `TransactionState` to `LookupInvoiceResponse` (https://github.com/rust-nostr/nostr/pull/1045)
 - Add `nip47::Method::as_str` method
 - Add `CommentTarget::{event, coordinate, external}` to point to a specific thing (https://github.com/rust-nostr/nostr/pull/1034)
 - Add `nips::nip73::Nip73Kind` and `TagStandard::Nip73Kind` (https://github.com/rust-nostr/nostr/pull/1039)